backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase>=2.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx>=0.26.0,<0.29.0
openai==1.3.5
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
python-dateutil==2.8.2
postgrest>=2.0.0
yfinance>=0.2.66

